Drop KSM metrics from shoot control-planes in the garden Prometheus
When the garden cluster is also a seed, it holds shoot control-planes.
This causes the VPA recommendation alerts from shoot control-planes to
fire twice: first, through the shoot Prometheus and, second, through the
garden Prometheus. This is however inconsistent when the garden cluster
is not a seed because in such cases it does not hold those namespaces.
This commit fixes this by adding a new `metric_relabel_config` for the
kube-state-metrics job in the garden Prometheus that drops metrics from
the shoot control-plane namespaces.