Adapt `required` controller
Earlier, the controller reacted on single extension events, i.e., when they were created during the `garden` creation flow. This is too late when the adjusted reconciliation flow should wait to all required garden extensions to be ready. This commit changes the `required` controller to consult the `Garden` resource to calculate required extensions instead.

The integration test is refactored into ordered `It`s along the way.