[GEP-28] Prepare `gardener/resourcemanager` component for "bootstrap control plane" case (#11343)
* Introduce `BootstrapControlPlaneNode` option in GRM values

- set hostNetwork=true
- tolerate all taints
- set API server address to 127.0.0.1

* Prefactor: Introduce ResponsibilityMode in GRM values

For better distinction which features are enabled.
This is a prefactoring step because we'll introduce a third mode `ForSourceAndTarget`.

* Introduce `ForSourceAndTarget` responsiblity mode

- enable certain webhooks

* Simply GRM component configuration

Instead of providing many arguments in the functions, just pass the
values that should be overridden with whatever user-desired config, and
apply them to the default values of the respective function.

* Add unit tests

* Address PR review feedback