Fix availability of Gardener `APIService`s (#11400)
* Fix availability of Gardener `APIService`s

Disable `endpoint` controller in virtual clusters. Esp. in the virtual garden cluster this caused issues because very time the KCM was rolled (or another leader was elected) the `endpoint` controller removed the IPs from the operator managed `gardener-apiserver` endpoint. As a consequence Gardener APIs were unreachable up to `1m`, until the `gardener-resource-manager` re-applied the endpoints object.

The `persistentvolumeclaim-protection` and `ephemeral-volume` controllers are disabled along the way. Also, the `discovery/v1` API group is now disabled by default for virtual (or workerless) clusters since we don't see any use-case for `EndpointSlice`s  for those.

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

* Address PR review feedback

---------

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>