Revisit the `VerticalPodAutoscalerCappedRecommendation` alert (#11325)
* Fix the alert query for a race condition

The original query assumes that querying Prometheus for the VPA state at
a certain time always yields a result consistent with the actual state
at that specific time. However, we found the KSM metrics are refreshed
independently of each other. A race condition can occur after a VPA
recommendation update if the Prometheus scrape happens at an instant
when the uncapped target has already been refreshed but the regular
target has not.

The proposed fix is to use the `count_over_time()` function to ensure
the uncapped target is larger than the regular target for 5m, and alert
only in such case. This is similar to adding `for: 5m` in the alert
definition, but it avoids the pending state, which would be incorrect.

* Do not count the VPAs for the garden cluster

Counting the VPAs in the garden cluster has the caveat that labels are
lost during aggregation. This, in turn, complicates debugging which VPA
and containers are affected. To keep the count and the labels, we would
need two queries, similarly as in the seed and shoot cases (i.e., the
aggregating query in those cases is the federated one). We opt to have
only one query for the sake of simplification and an alert per VPA in
the garden cluster.

* Improve the aggregated VPA capped recommendation alert names

This adds the "Count" suffix to the aggregated alert names. The alert
text is also improved by providing more indications of what to do when
receiving the alert.

* Drop KSM metrics from shoot control-planes in the garden Prometheus

When the garden cluster is also a seed, it holds shoot control-planes.
This causes the VPA recommendation alerts from shoot control-planes to
fire twice: first, through the shoot Prometheus and, second, through the
garden Prometheus. This is however inconsistent when the garden cluster
is not a seed because in such cases it does not hold those namespaces.
This commit fixes this by adding a new `metric_relabel_config` for the
kube-state-metrics job in the garden Prometheus that drops metrics from
the shoot control-plane namespaces.