[operator] Ensure Garden environment is clean when `Garden` is deleted (#11460)
* Cleanup garbage-collectable resources after `Garden` deletion

Since GRM is gone, the garbage-collector (which would normally clean
them up) is gone, so let's do it ourselves (via gardener-operator) at
the very end of the `Garden` deletion flow.

* Wait for `Garden` to be reconciled AND HEALTHY

Previously, we only waited for it to be reconciled, but it would
actually be better if we also wait for it to be healthy before we make
assertions about its state.
We do the same for Shoot e2e tests.

* Do not forget to destroy plutono dashboard `ConfigMap`

* [provider-local] Delete generated `BackupBucket` secret