Change `prometheus` label of `vpa` `ServiceMonitors` from `seed` to `garden` when deployed by `garden operator`  (#11318)
* feat/vpa: add `IsGardenCluster` field to `Values` configuration

This commit adds a new `IsGardenCluster` field to the `VPA`
values configuration. The purpose of the field is to indicate
if the cluster, at which the VPA is getting deployed, is
registered as a `Garden` cluster.

Also:
- Based on the new `IsGardenCluster` field, make the
  `getPrometheusLabel()` function return `garden` label.

* feat/controller/garden: instantiate VPA with `IsGardenCluster` config param

This commit instruments the `garden` controller to utilize the
`IsGardenCluster` parameter when instantiating new VPA entity.

* feat/controller/seed: instantiate `VPA` with `IsGardenCluster` config param

This commit instruments the `seed` controller to instantiate `VPA`
entries with `IsGardenCluster` config param.

* feat/vpa/test: add unit tests for `seed`/`garden` `ServiceMonitor` labels

This commit adds unit tests for VPA `ServiceMonitor` labels
created based on the the different configuration parameters
with which the VPA was instantiated.

Also:
- Adds a helper `vpaFor` function to the `vpa_test.go` to be
  used for instantiating VPAs for the different cluster types.

* feat/vpa/test: add unit tests for `shoot` cluster `ServiceMonitor` labels

This commit adds unit tests that verigy `shoot` cluster `VPA`

* `admission-controller`
* `recommender`

have their `prometheus` labels set to `shoot` *no matter* if the
`IsGardenCluster` configuration flag is `true` or `false`.

* chore/vpa/test: rename `ServiceMonitor` vars to match `actual` and `expected` contexts

* chore/garden/components: remove redundant `IsGardenCluster` var

This commit removes a redundant variable storing the flag
indicating if the cluster is `Garden` or not.

* chore/vpa/test: remove redundant empty lines

This commit remove leftover empty lines from the `vpa_test.go`.

* chore/vpa/test: Use `vpaFor` in `BeforeEach` for `seed` cluster type

This commit makes use of the `vpaFor` utility in the `BeforeEach`
function within the `Seed` cluster type context.

Also:
- Improve the wording for the test cases that handle `prometheus` labels
  in the context of a `seed` cluster type.

* chore/vpa/test: group `shoot` ServiceMonitor label tests by `IsGardenCluster` value

This commit refactors the `vpa-recommender`/`vpa-admission-controller`
ServiceMonitor labels tests by grouping the cases with respect to the
`IsGardenCluster` value.

Also:
- Improve the wording in the contexts and extract commonly used
  variables in shared `BeforeEach` functions.
- Remove redundant `IsGardenCluster` variables and call `vpaFor` with
  the `bool` value directly.

* feat/vpa/recommender: add `NetworkPolicy` for scraping `garden` targets

This commit adds logic that switches b/w `garden` and `seed` targets for
scraping `vpa-recommender` metrics.

* feat/vpa/admissioncontroller: add `NetworkPolicy` for scraping `garden` targets

This commit instruments the `vpa-admission-controller` to configure it's
`NetworkPolicy` records based on the passed configuration.

* chore/vpa/test: update `shoot` vpa test labels to match `seed` values

This commit updates the `shoot` vpa test labels to match the ones
defined within the `seed` contexts.

Also:
- Rename expected `serviceMonitorExpected` variable name to
  `serviceMonitorRecommender` to be consistent with existing usages in
  other test scenarios.
- Remove redundant validation of `expected` and `actual` labels as the
  `Expect(<expected>).To(Equal(<actual>))` will compare the objects.
- Update `varFor` function's `isGardenCluster` variable name to start with
  lowercase char.

* chore/vpa/test: add `isGardenCluster` argument to `serviceMonitorRecommenderFor` utility

This commit adds a new `isGardenCluster` argument to the `serviceMonitorRecommenderFor`
utility function in order to reduce the manual changes of objects within
the tests.

It also updates the existing usages of the function to include the newly
added argument.

* feat/vpa/test: add new `serviceMonitorAdmissionControllerFor` utility

This commit adds a new utility function `serviceMonitorAdmissionControllerFor`
that is used for creating new `vpa-admission-controller` `ServiceMonitor`
records in the context of the `ClusterType` and `isGardenCluster` flag.

Also:
- Update all usages of the current `serviceMonitorAdmissionController`
  variable with the newly added utility function.

* feat/vpa/test: check correct network policy annotation for `vpa-recommender`

This commit adds a new `isGardenCluster` flag to the
`serviceRecommenderFor()` utility to update the records in the context
of a `garden` cluster.

Also:
- Updates the existing usages of the utility.
- Adds a new test to check if the `vpa-recommender` service is annotated
  with the correct network policy annotation.

* feat/vpa/test: check correct network policy annotation for `vpa-admission-controller`

This commit adds a new `isGardenCluster` argument to the
`serviceAdmissionControllerFor()` utility. The new flag is used
for including additional `NetworkPolicy` annotation that enables
metrics scraping from the service.

Also:
- Update the existing usages of the utility.
- Add unit tests to validate that the correct annotation is being set.

* chore/vpa/test: match `shoot` and `seed` ServiceMonitor test titles

This commit matches the `shoot` and `seed` ServiceMonitor test titles
so that they all follow the same naming convention.

* chore/vpa/test: remove implicit use of `ResourceVersion` in admission controller utility

This commit removes the implicit setting of `.ResourceVersion` field in
`serviceMonitorAdmissionControllerFor()` utility. Since all of the
`*For()` utilities do not set the field, but rely on the func caller to
set it, this change updates the utiilty to match the current pattern.

Also:
- Reorder the variable names in the `shoot` cluster tests and update
  their names to match `expected` and `actual` notation.