Allow to configure `hostNetwork`, `replicas` and `tolerations` for controller installations.
In the autonomous shoot cluster scenario, some controller installations
may need to run in the host network. They may also need to be scaled
down in a single node scenario or may need to tolerate additional taints
during bootstrap.
This change adds the corresponding parameters as optional attributes to
the controller installation reconciler.