Allow to configure `hostNetwork`, `replicas`, `tolerations` and usable ports for controller installations.
In the autonomous shoot cluster scenario, some controller installations
may need to run in the host network. They may also need to be scaled
down in a single node scenario or may need to tolerate additional taints
during bootstrap. Running multiple controller installations in the host
network also requires managing the usable ports so that two controllers
do not clash in their port usage.
This change adds the one flag to the controller installation reconciler
that will trigger the corresponding changes.
Notably, the handling of the usable ports is focussed on the `gardenadm`
scenario and is explicitly not supposed to be run as a controller.