Add `VPA` dashboards to `plutono` running in a `Garden` cluster with enabled VPA (#11391)
* feat/plutono: add `vpa` dashboards to `garden` cluster plutono instance

This commit adds the `vpa` dashboards to the `plutono` instantiation
when deployed in a `garden` cluster context.

Since the `vpa-admission-controller` and `vpa-recommender` metrics are
now being collected as part of the `garden` prometheus instance,
we need the `vpa` dashboards to display the collected metrics.

* feat/controller/garden: use `vpaEnable()` when instantiating new `plutono`

This commit instruments the `newPlutono` method to start using the
`vpaEnabled()` to calculate the VPA enable flag instead of the
hard-coded `false` value.

* feat/plutono/test: validate dashboards count when garden cluster VPA is enabled

This commit adds a unit test that validates the correct number of
dashboards configued in plutono if the cluster is a `garden` and
`vpa` is enabled.

* chore/plutono: reuse existing common dashboards embed for VPA in Garden

This commit instruments the plutono dashboard loading to reuse the
existing `commonDashboards` embed when picking the `vpa` dashboards
when in Garden cluster and `VPAEnabled=true`