Consider `status.seedName` instead of `spec.seedName` (#11521)
* shoot care: consider `status.seedName` instead of `spec.seedName`

* seed lifecycle: consider `status.seedName` instead of `spec.seedName`

The integration test is improved to allow focusing on specific `It`s by creating the `garden` namespace in the general `JustBeforeEach`
instead of only for the first `It` that needed it.

* shoot migration: use strategic merge patch

This patch changes constraints, which should not be changed using a patch without optimistic locking.
Switch to strategic merge patch to ensure this doesn't accidentally remove constraints.

* shoot: fix copy paste mistake