fix: SSH tunnel livenessProbe no longer fails with slow DNS (#11538)
The netstat command in the livenessProbe of the SSH tunnel for local development sometimes failed when the DNS resolution takes more than 1 second. The `-n` parameter skips name resolution. `grep` can then no longer look for the local hostname, so adding the `-p` parameter adds info about the process name, which we can grep for instead.