Use service discovery to federate to the seed Prometheus
If the scrape URL is provided using static configurations, Prometheus v2
will try to add the port based on the protocol, while Prometheus v3 will
not. This is an issue with the seed Prometheus where the "instance"
label changes from "prometheus-cache.garden.svc:80" to
"prometheus-cache.garden.svc" after the upgrade.

This commit updates the seed Prometheus to use service discovery to
federate from the cache Prometheus in the seed. This not only resolves
the label mismatch but will also make federation configurations
consistent in the following commits.