Use service discovery to federate to the longterm Prometheus
The longterm Prometheus is also affected by the label change when
federating from the garden Prometheus. In this case, the "instance" label
changes from "prometheus-garden:80" to "prometheus-garden".

Using service discovery will update the "instance" label to
"prometheus-garden.garden.svc:80", but it will also make it consistent
with other federation configurations.