Upgrade Prometheus to v3.2 (#11552)
* Update Prometheus to v3.2.1

* Fix Prometheus rule tests in etcd

The unit tests were originally misconfigured because they query a test
time series that only exists for 10 minutes and gets staled for 5
additional minutes. Note such time series is defined as 0x20 but the
evaluation interval is 30 seconds. Furthermore, the tests would check
the alert after 15 minutes despite it triggers after 10 minutes. It
seems, in this situation, the alert remains triggered after 15 minutes
in Prometheus v2 but not in Prometheus v3.

This commit adjusts the tests to evaluate the alert after 10 minutes, in
line with its configuration. This resolves the issue in Prometheus v3
where the alert does not remain triggered after 15 minutes.

* Use service discovery to federate to the seed Prometheus

If the scrape URL is provided using static configurations, Prometheus v2
will try to add the port based on the protocol, while Prometheus v3 will
not. This is an issue with the seed Prometheus where the "instance"
label changes from "prometheus-cache.garden.svc:80" to
"prometheus-cache.garden.svc" after the upgrade.

This commit updates the seed Prometheus to use service discovery to
federate from the cache Prometheus in the seed. This not only resolves
the label mismatch but will also make federation configurations
consistent in the following commits.

* Use service discovery to federate to the shoot Prometheus

This commit is analog to previous commit, but for the shoot Prometheus.

* Use service discovery to federate to the longterm Prometheus

The longterm Prometheus is also affected by the label change when
federating from the garden Prometheus. In this case, the "instance" label
changes from "prometheus-garden:80" to "prometheus-garden".

Using service discovery will update the "instance" label to
"prometheus-garden.garden.svc:80", but it will also make it consistent
with other federation configurations.