Write seed-specific `Shoot` endpoint served with the control plane wildcard certificate into `.status.advertisedAddresses` (#11612)
* Documentation

* Augment advertised addresses with wildcard TLS endpoint

* Generate kubeconfigs with `Cluster` using wildcard TLS endpoint

* Do not default the server CA to the signing CA of the certificate

This code only works under the assumption that the signing CA of the
certificate is the same as the one for the server.

This was introduced 3y ago with
https://github.com/gardener/gardener/pull/5524/commits/039631fb7fe79398d18882f4810a01d949d6016b
when this was the case for Shoots.

Since a longer time, however, we have introduced a dedicated client CA
for shoots, i.e., this defaulting never works in our case.
Let's remove it and rely on the CA data being properly provided via the
`KubeConfigRequest` struct.

* Address PR review feedback