Add `/ko-app` prefix for GNA when build locally
For regular shoots, provider-local's webhook on OperatingSystemConfig
injects the `/ko-app` prefix. However, in the gardenadm scenario, we
don't have provider-local's webhook, so the only feasible option is to
handle it ourselves. Not nice, but perhaps the most pragmatic solution.

We might have to do the same for the gardener-node-init script later
when we implement `gardenadm` join (then we can drop the webhook in
provider-local for good).