Patch `ShootState` integration tests addressing flakes when managing resource finalizer  (#11790)
* chore/shootstate/tests: run add/remove finalizer within `Eventually` block

This commit instruments the add/remove finalizer logic to run within an
`Eventually` block so that the `ShootState` resource version gets
updated in teh `testClient` cache and the patch is applied to the
latest object version.

* chore/shootstate/tests: update `DeferCleanUp` resource removal order

This commit updates the order at which the resources are being cleanup
in the `DeferCleanUp` function.

* chore/shootstate/test: move `addFinalizer` helper func outside of `Describe` block

This commit moves the `addFinalizer` helper function outside of the
`Describe` block as it's an external utility for the tests.

* Cosmetic nit

---------

Co-authored-by: Ismail Alidzhikov <9372594+ialidzhikov@users.noreply.github.com>