Dry-run provider secret creation to ensure it is sane
This way, we ensure that all `*Binding` resources can be considered
"safe to be used" UNDER THE CONDITION that the provider type of the
`*Binding` matches with the provider type of the target resource (e.g.,
a `Shoot` with provider type `foo` must ensure that it only references
a `*Binding` with provider type `foo`).

This is already ensured by https://github.com/gardener/gardener/blob/2a9c566f89a6a16e654193945250f60755c560b9/plugin/pkg/shoot/validator/admission.go#L1073-L1091.
As of today, the `Shoot` API is the only one using `*Binding` resources.

Note that we don't need to do this dry-run-create for
`WorkloadIdentity`s because we already ensure that the provider types of
it and the `CredentialsBinding` match (see previous commit). This way,
we ensure that a `WorkloadIdentity` must have already successfully
passed the provider extension check when referenced in
a `CredentialsBinding` with the same type.

Co-Authored-By: Tim Usner <tim.usner@sap.com>