Allow the Gardenlet to see the ManagedSeed object of its own cluster
In `pkg/gardenlet/controller/seed/seed/reconciler.go`

    (r *Reconciler) Reconcile(...) ...

the gardenlet tries to detect if it is running in a managed seed:

    managedSeed, err := kubernetesutils.GetManagedSeedByName(ctx, r.GardenClient, seed.Name)
    if err != nil {
      if !apierrors.IsNotFound(err) {
        return reconcile.Result{}, fmt.Errorf("failed to get managed seed: %w", err)
      }
    }
    isManagedSeed := managedSeed != nil

by getting the ManagedSeed object for its own name.
If it finds a ManagedSeed object, then it knows that it is running in a Gardener managed seed cluster.

This information, whether the gardenlet is running in a managed seed or not,
is used for this older PR: https://github.com/gardener/gardener/pull/9716

If it is a managed seed, the CA of the kubelet's certificate is not issued by
the Kubernetes cluster's root CA (see details in that PR) and hence the TLS
verification should be skipped in the cadvisor/kubelet scrape job of the cache
Prometheus.

The PR https://github.com/gardener/gardener/pull/11773/commits/d19cb61f7bb1a93e40941b9ea25ecb94335f4026
that was released first with v1.117.0 introduced a change,
such that the gardenlet can no longer see the ManagedSeed object of its own cluster anymore.

This breaks the heuristic above to detect if the gardenlet is running in a managed seed or not.

Conceptually it is valid that the gardenlet should be able to see the ManagedSeed object of its own cluster.
Currently this is controlled by the

    name.seed.gardener.cloud/<seed-name>: "true"

labels on the ManagedSeed object using admission webhooks.

By adding a new label to the ManagedSeed object for the gardenlet's own cluster,
the gardenlet will be able to see the ManagedSeed object again, not only for its seed,
but also for its own cluster.

We confirmed this fix in a managed seed in a local setup:

Create a shoot and upgrade it to a managed seed and add an `/etc/hosts` entry
for the managed seed's API
`api.managedseed.garden.external.local.gardener.cloud`:

    k apply -f example/provider-local/managedseeds/shoot-managedseed.yaml
    k apply -f example/provider-local/managedseeds/managedseed.yaml

In the managed seed's garden namespace:

    k port-forward prometheus-cache-0 9090
    curl -sS localhost:9090/api/v1/status/config | jq .data.yaml -r | grep "job_name: cadvisor" -A 20 | grep skip
    --> verify the expected Prometheus configuration: TLS verification should be skipped

In the virtual Garden cluster:

    k get managedseeds -n garden managedseed -o yaml | yq .metadata.labels | grep name.seed.gardener.cloud
    --> verify the new label on the ManagedSeed object

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Victor Herrero Otal <victor.herrero.otal@sap.com>