Remove the superfluous label from the managed seeds
PR #11965 attempted to address a regression in the monitoring stack but required
manual mutation and reconciliation of `ManagedSeed` and `Seed` objects after
deployment.

In some Gardener landscapes, `ManagedSeed` objects were manually reconciled to
add the label via the mutating webhook:

    name.seed.gardener.cloud/<seed-name>: "true"

This allowed the gardenlet to access its own `ManagedSeed` object and fixed the
monitoring regression.

However, this PR adopts a different approach, eliminating the need for the
gardenlet to access its own `ManagedSeed`.

This commit removes the no longer necessary label from the `ManagedSeed` objects
that might be present in some installations. The gardener-controller-manager
sends an empty patch during its startup phase to the `ManagedSeed` resource. The
mutating webhook (that was changed in the previous commit) can then delete the
superfluous label.

Without this cleanup, the gardenlet would continue attempting to reconcile its
own `ManagedSeed` object, resulting in error logs, until the `ManagedSeed`
object is mutated for some other reason.

This snippet is added to an existing block of code that also deals with seed
labels and that is scheduled to be deleted after Gardener v1.119 was released.
The scheduled removal after v1.119 is just fine for this scenario as well:

- the regression was introduced in v1.117
- this PR will be regularly released with Gardener v1.119
  - it will be backported to v1.117.x and v1.118.x

So it is fine to remove this code after v1.119.

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Victor Herrero Otal <victor.herrero.otal@sap.com>