[GEP-31] Controller optimizations and Add e2e tests (#11953)
* Enable featuregate locally

* [node-agent] Improve OS version handling

* [node-agent] Handle failed case

When the OSC secret is updated, do not wait for the ReadyForUpdate condition if UpdateFailed is already present, because the node is already drained, we can retry the update

* [node-agent] Do not label node if it's already labelled

* -----------------Empty separator commit----------------

* [provider-local] Set machine image as in-place supported in cloudprofile local

* [provider-local] Add OS update command to provider-local

* [provider-local] Add pool name in machine-deployment

* -----------------Empty separator commit----------------

* [gardenlet] Improve shoot status reconciler

Trigger a reconciliation when all manual in-place workers are updated and credential rotation status is preparing

* [gardenlet] Use optimistic lock for updating shoot status

* [gardenlet] Set optimistic lock to false for existing updates

* -----------------Empty separator commit----------------

* Do not set enableStaticTokenKubeconfig field

* Improve predicate in worker controller

MCM can be slow in syncing node conditions to the machine, so it could happen that the SelectedForUpdate is missed and ReadyForUpdate condition is added. To circumvent this limitation, the predicate is adapted.

* -----------------Empty separator commit----------------

* Add e2e test for credential rotation

Extract common logic into helper function for reuse

* Add e2e tests for update

* Make linter happy

* Improve validation for update strategy field

Validate feature gate only on creation. We should not break existing shoots if we decide to turn off the feature gate.

* Do not deny existing worker pools with in-place strategy

* Address PR review feedback from acumino

* Address PR review feedback from rfranzke

* Rewrite os-release in worker controller

* Update skaffold dependencies