Fall back to `/ko-app` in OSC `ImageRef`
We keep the `/ko-app` prefix in the local OSC file for one more release.
Without this, the upgrade tests fail because the old node-agent on existing nodes will not
be able to extract the node-agent binary from the binary without the `/ko-app` prefix in
the `filePathInImage` field.
We need to perform this clean up in two steps over two releases.