Check if the seed is a shoot via the kube-system/shoot-info ConfigMap
Since PR #11773, the gardenlet can no longer access its own ManagedSeed resource
due to a label selector applied at the manager level.

Instead of reintroducing visibility for the ManagedSeed resource, as attempted
in PR #11965 (and now reverted), we rely on the presence of the
`kube-system/shoot-info` ConfigMap to determine if the seed is a Gardener shoot.

The Kubelet scrape configuration should skip TLS verification if the seed is a
Gardener shoot, as the kubelet's certificate in Gardener shoots is issued by a
separate Kubelet CA.

The original check for the ManagedSeed was not entirely accurate, as it is
irrelevant how the gardenlet is deployed to a Kubernetes cluster. What matters
is whether the cluster is an IaaS Kubernetes cluster (where the kubelet's
certificate is issued by the Kubernetes CA) or a Gardener shoot cluster (where
the kubelet's certificate is issued by a separate Kubelet CA).

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Victor Herrero Otal <victor.herrero.otal@sap.com>