Use `/livez` endpoint of `kube-proxy` instead of `/healthz` endpoint for probes. (#12015)
* Use `/livez` endpoint of `kube-proxy` instead of `/healthz` endpoint for
probes.

`kube-proxy` serves two paths on the health endpoint (port 10256) `/livez` and `/healthz`.
Both check whether `kube-proxy` is healthy, but the latter also takes
the node into account. It reports failure if the node either has a
deletion timestamp or a taint by `cluster-autoscaler` indicating that
the node is about to be deleted.
Previously, we used the `/healthz` endpoint for the readiness probe,
which worked fine for indicating that `kube-proxy` as critical component
was ready. However, it could lead to prolonged readiness failures in
case a node was about to be deleted by `cluster-autoscaler`, but some
`PodDisruptionBudget` or `terminationGracePeriod` delayed the removal.
The result was failures in the system components health check.
Now, these failures should not be present anymore.

* Address Review Feedback