Use technical ID as namespace when not running control plane
We want to reuse the botanist code for `gardenadm bootstrap` as well.
This botanist instance also manages autonomous shoots but it doesn't run the control plane.
Hence, the deployment mode (e.g., for gardener-resource-manager) is different than in the resulting
autonomous shoot cluster itself.
That's why we distinguish between `Shoot.IsAutonomous()` and `Shoot.RunsControlPlane()`.