[GEP-28] `gardenadm bootstrap`: Deploy gardener-resource-manager and provider extension (#12052)
* Prefactor: extract `extensionsv1alpha1.AllExtensionKinds`

* Use technical ID as namespace when not running control plane

We want to reuse the botanist code for `gardenadm bootstrap` as well.
This botanist instance also manages autonomous shoots but it doesn't run the control plane.
Hence, the deployment mode (e.g., for gardener-resource-manager) is different than in the resulting
autonomous shoot cluster itself.
That's why we distinguish between `Shoot.IsAutonomous()` and `Shoot.RunsControlPlane()`.

* Filter extension kinds for autonomous botanist

With this, we only deploy the provider extension (`Infrastructure` and `Worker`) in `gardenadm bootstrap`.

The TODO about distinguishing medium-touch and high-touch is basically moved from the generic botanist package (`shoot.go`)
to the autonomous botanist package (`extensions.go`).

* Deploy gardener-resource-manager and provider extension

* Fix context handling in e2e tests, extend timeout

* Address review suggestions