Drop deprecated Istio metric
The Istio Mesh dashboard is based on the upstream dashboards. As pointed
out in

    https://github.com/istio/istio/issues/31516

the old

    istio_request_duration_seconds_bucket

metric was generated by a (deprecated) component called Istio Mixer. The new

    istio_request_duration_milliseconds_bucket

metric is generated by the Istio Proxy that's still active.

The Istio Mesh upstream dashboards can be downloaded from

    https://grafana.com/grafana/dashboards/7639-istio-mesh-dashboard/

to validate the most recent versions do not reference the old metric
anymore.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>