Drop Istio histogram metrics
The Istio metrics:

    istio_request_duration_milliseconds_bucket
    istio_response_bytes_bucket
    istio_request_bytes_bucket

generate a significant amount of data. We've observed instances where
these metrics can consume up to 45% of a 2GB / 54h Prometheus block.
This poses a risk of triggering the 15GB disk retention limit of
Prometheus sooner than the intended 30-day retention period, potentially
hindering our ability to debug historical issues. We have decided
against increasing the disk size for now because the cost of these
metrics consuming 45% of the storage is excessive. First, we should
optimize these metrics to ensure efficient use of resources.

This commit removes the bucket submetric for these histograms, but makes
sure to keep the sum and count submetrics, which do not use much
storage. However, the capability to calculate percentiles is lost, but
we can still measure the average. The submetrics

    istio_response_bytes_sum
    istio_response_bytes_count
    istio_request_bytes_sum
    istio_request_bytes_count

are not used in Gardener but still retained for interactive queries
where the sum or count might be useful for debugging purposes. The
submetrics

    istio_request_duration_milliseconds_sum
    istio_request_duration_milliseconds_count

are used in a dashboard to display the average latency as a replacement
to the 50th, 90th and 99th percentiles, which is not possible anymore as
the bucket submetric is dropped.

If using only the average turns out to be insufficient, we may need to
deploy a dedicated Istio Prometheus instance to compute aggregated
recording rules and federate from it:

https://istio.io/latest/docs/ops/best-practices/observability/

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>