Remove defaulting and setting `maxEmptyBulkDelete`
If the value is not explicitly set by the user, don't fall back
to some other value since that causes issues with validation.
It's difficult to figure out when the user actually sets the value
in that case. If the value is set, use that to set maxScaleDownParallelism
and only use the latter when launching cluster autoscaler.