Support `gardenadm join`
Machine name must be set correctly to make this work. Additionally, the last-applied-osc.yaml file must not be written when GNA Reconciler applies the init OSC. Otherwise, it would delete the `machine-name` file right away because it is in the init OSC but not in the original one.
While gardener-node-agent is bootstrapping there is no node in the `gardenadm join` scenario. Thus, it requires access to all OSC secrets because node-agent-authorizer does not know which worker pool the new GNA is using. If there would be one OSC per node, we could restrict the access to the OSC secrets which are not used on any node.