Clean up obsolete "prometheus-" folder from Prometheus volumes
In GEP-19, the Prometheus operator was introduced in Gardener to manage
Prometheus instances. This operator configures Prometheus statefulsets
to use the "prometheus-db" folder as a subpath for storage. Previously,
the "prometheus-" folder was used instead, as in:

    https://github.com/gardener/gardener/blob/96335a6487944f05322ac8f182f46c1e738ee2ff/pkg/component/monitoring/charts/seed-monitoring/charts/core/charts/prometheus/templates/prometheus.yaml#L134-L135

During the migration to the Prometheus operator, there was a disk
migration to move data between the two folders.

However, the old "prometheus-" folder still exists in some volumes,
consuming up to 7GB of data in certain control-plane Prometheus
instances. This poses a risk of Prometheus running out of disk space, as
it operates on a 20GB volume with a disk retention setting of 15GB.

To address this, this commit introduces an init container to Prometheus
to remove the obsolete "prometheus-" folder from the volume.
Unfortunately, identifying the exact sequence of steps or potential disk
migration failures that led to this situation is challenging after this
time and will remain unresolved.