Clean up obsolete "prometheus-" folder from Prometheus volumes
In GEP-19, the Prometheus operator was introduced in Gardener to manage
Prometheus instances. This operator configures Prometheus statefulsets
to use the "prometheus-db" folder as a subpath for storage:

    > k get pods prometheus-shoot-0 -o json | jq -r '
        (.spec.volumes[] | select(.name == "prometheus-db")),
        (.spec.containers[].volumeMounts[] | select(.name == "prometheus-db"))'
    {
      "name": "prometheus-db",
      "persistentVolumeClaim": {
        "claimName": "prometheus-db-prometheus-shoot-0"
      }
    }
    {
      "mountPath": "/prometheus",
      "name": "prometheus-db",
      "subPath": "prometheus-db"
    }

Previously, the "prometheus-" folder was used instead, as in:

    https://github.com/gardener/gardener/blob/96335a6487944f05322ac8f182f46c1e738ee2ff/pkg/component/monitoring/charts/seed-monitoring/charts/core/charts/prometheus/templates/prometheus.yaml#L134-L135

During the migration to the Prometheus operator, there was a disk
migration to move data between the two folders:

    https://github.com/gardener/gardener/pull/9128/commits/ca0a1863fca24146d6689bdb6b8d9017a44441d9#diff-c6e1ff35f50162ec0f1[…]1c0d5253120d9c71R101-R106

However, the old "prometheus-" folder still exists in some volumes,
consuming up to 7GB of data in certain control-plane Prometheus
instances. This poses a risk of Prometheus running out of disk space, as
it operates on a 20GB volume with a disk retention setting of 15GB.

To address this, this commit introduces an init container to Prometheus
to remove the obsolete "prometheus-" folder from the volume.
Unfortunately, identifying the exact sequence of steps or potential disk
migration failures that led to this situation is challenging after this
time and will remain unresolved.